DTO 수정, 공지게시판 글쓰기 기능 구현

+++ finalproject6/src/main/java/com/kh/final6/entity/StoreLikeDto.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class StoreLikeDto {
 	private long likeNo;
 	private long storeNo;
-	private long memberNo;
+	private int memberNo;
 	private int likeCheck;
 }
